+++ __tests__/components/Footer.test.tsx
@@ -1,4 +1,4 @@
-import { Footer } from '@/components/Footer/Footer';
+import Footer from '@/components/Footer/';
 import { render } from '@testing-library/react';
 
 describe('Footer', () => {
@@ -9,7 +9,10 @@ describe('Footer', () => {
 
     it('should have a link to the repo', () => {
         const { container } = render(<Footer />);
-        expect(container.querySelector('a')).toHaveAttribute('href', '#');
+        expect(container.querySelector('a')).toHaveAttribute(
+            'href',
+            'https://github.com/Real-Dev-Squad/tiny-site-frontend'
+        );
     });
 
     it('should have a link to the repo', () => {
